% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for_datasets.R
\name{download_dataset}
\alias{download_dataset}
\title{Download dataset}
\usage{
download_dataset(urllink, filepath = NULL)
}
\arguments{
\item{urllink}{Download link/url}

\item{filepath}{If specified, the file will be downloaded to the specified location. If unspecified, it will be saved in the tmp directory}
}
\description{
Given a download link, obtained by using either `search_for_datasets` or `get_datasets_from_a_catalog`, this function will download the file.
}

