% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fetch_data}
\alias{fetch_data}
\title{Load data from the Government of India API.}
\usage{
fetch_data(res_id, api_key)
}
\arguments{
\item{res_id}{a string, JSON data resource id}

\item{api_key}{a string, private api key of user, can be obtained by signing up on \url{data.gov.in} portal}
}
\value{
data a data.frame, data from the Government of India API
}
\description{
\code{fetch_data} is the main function from this package to load the entire data set from the Government of India API.
}
\examples{
\dontrun{
### fetch a dataset using it's resource id and your personal API key
fetch_data(res_id = <res_id>, api_key = <api_key>)
}
}
\keyword{Name}

