% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{get_data}
\alias{get_data}
\title{Get data from the JSON data object}
\usage{
get_data(x)
}
\arguments{
\item{x}{a list, i.e. a JSON data object}
}
\value{
data a list, data from the JSON data object
}
\description{
This will return the data from the JSON data object.
}
\examples{
\dontrun{
###Return data from a JSON data object (obtained using get_JSON_doc())
get_data(x = JSON_doc)
}
}
\keyword{Name}

