% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{get_field_names}
\alias{get_field_names}
\title{Get field/variable names from the JSON data object}
\usage{
get_field_names(x)
}
\arguments{
\item{x}{a list, i.e. a JSON data object}
}
\value{
field_names a vector/list, of field names for JSON data object
}
\description{
This will return field names from the JSON data object.
}
\examples{
\dontrun{
###Return field names from a JSON data object (obtained using get_JSON_doc())
get_field_names(x = JSON_doc)
}
}
\keyword{Name}

