% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{get_field_type}
\alias{get_field_type}
\title{Get field/variable types from the JSON data object}
\usage{
get_field_type(x)
}
\arguments{
\item{x}{a list, i.e. a JSON data object}
}
\value{
field_types a list/vector, field type of each of the fields
}
\description{
This will return field types from the JSON data object.
}
\examples{
\dontrun{
###Return field types from a JSON data object (obtained using get_JSON_doc())
get_field_names(x = JSON_doc)
}
}
\keyword{Name}
