% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ogdindia_api_key}
\alias{ogdindia_api_key}
\title{Get or set OGDINDIA_API_KEY value}
\usage{
ogdindia_api_key(force = FALSE)
}
\arguments{
\item{force}{Force setting a new PassiveTotal API key for the current environment?}
}
\value{
atomic character vector containing the Open Government Data India API key
}
\description{
The API wrapper functions in this package all rely on a Open Government Data India API
key residing in the environment variable \code{OGDINDIA_API_KEY}. The
easiest way to accomplish this is to set it in the `.Renviron` file in your
home directory.
}
