% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ogdindiar.R
\docType{package}
\name{ogdindiar}
\alias{ogdindiar}
\alias{ogdindiar-package}
\title{ogdindiar: Provides API access to selected datasets on Open Government Data - India Portal.}
\description{
The ogdindiar package provides two categories of important functions:
Downloading entire datasets,
Downloading specific elelments based on certain conditions.
}
\section{ogdindiar functions}{

fetch_data
}

