% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rectify_field_type}
\alias{rectify_field_type}
\title{Apply field type correction based on accompanied metadata}
\usage{
rectify_field_type(d_in, d_fields)
}
\arguments{
\item{d_in, }{a data.frame on which the correction is to be applied.}

\item{d_fields, }{a data.frame containing fields metadata}
}
\value{
data corrected data.frame
}
\description{
\code{rectify_field_type} will convert select fields to numeric based on accompanied metadata
}
\examples{
\dontrun{
rectify_field_type(data_stage2, data_field_type)
}
}
\keyword{Internal,}
\keyword{rectify}
