% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{to_data_frame}
\alias{to_data_frame}
\title{Convert data from list to a data.frame}
\usage{
to_data_frame(lst_elmnt)
}
\arguments{
\item{lst_elmnt}{a list of data from a JSON data object}
}
\value{
data a data.frame, data from the JSON data object
}
\description{
\code{to_data_frame} will convert data from 'list' to a 'data.frame'.
}
\examples{
\dontrun{
###Convert a list to data.frame
to_data_frame(x = get_data(JSON_list))
}
}
\keyword{Name}
