% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fetch_data}
\alias{fetch_data}
\title{Load data from the Government of India API.}
\usage{
fetch_data(
  res_id,
  filter = NULL,
  select = NULL,
  sort = NULL,
  field_type_correction = TRUE,
  max_obs = 500
)
}
\arguments{
\item{res_id}{a string, JSON data resource id}

\item{filter}{a named vector, specifying equality constrainsts of the form "variable" = "condition"}

\item{select}{a vector, specifying variables/fields to be selected}

\item{sort}{a named vector, specifying sort order in the form "variable" = "order"}

\item{field_type_correction}{boolean, whether to apply field type correction. All data fields are downloaded as character and then corrected (if at all) based on accompanying metadata}

\item{max_obs}{an integer, specifying maximum no of observations to fetch (will be rounded UP to the nearest 100)}
}
\value{
list a list of 2 elements - data from the Government of India API, and metadata, additional information about the fields
}
\description{
\code{fetch_data} is the main function from this package to load the entire data set from the Government of India API.
}
\examples{
\dontrun{
### fetch a dataset using it's resource id and your personal API key
# Basic Use:
fetch_data(res_id = "60a68cec-7d1a-4e0e-a7eb-73ee1c7f29b7")

# Advanced Use, specifying additional parameters
fetch_data(res_id = "60a68cec-7d1a-4e0e-a7eb-73ee1c7f29b7"
           filter = c("state" = "Maharashtra"), 
           select = c("s_no_","constituency","state"),
           sort = c("s_no_" = "asc","constituency" = "desc"))
}
}
\keyword{Name}
