% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for_datasets.R
\name{get_datasets_from_a_catalog}
\alias{get_datasets_from_a_catalog}
\title{get data sets for a catalog}
\usage{
get_datasets_from_a_catalog(catalog_link, limit_pages = 5L)
}
\arguments{
\item{catalog_link}{Link to the catalog}

\item{limit_pages}{Limit the number of pages that the function should request. Each page contains a list of data sets.}
}
\description{
Get the list of data sets and related info for a catalog
}
\examples{
\dontrun{
get_datasets_from_a_catalog(
'https://data.gov.in/catalog/fishing-harbours-fisheries-statistics-2014',
limit_pages = Inf)
}
}
\seealso{
search_for_datasets
}

