% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_JSON_doc}
\alias{get_JSON_doc}
\title{Get JSON data for requested data resource}
\usage{
get_JSON_doc(link = "https://data.gov.in/api/datastore/resource.json?",
  res_id, offset, no_elements, filter, select, sort, verbose = FALSE)
}
\arguments{
\item{link}{a string, general JSON data link}

\item{res_id}{a string, JSON data resource id}

\item{offset}{an integer, offset of 1 corresponds to 100 elements}

\item{no_elements}{an integer, no of elements to download a value between 1 to 100}

\item{filter}{a named vector, specifying equality constrainsts of the form "variable" = "condition"}

\item{select}{a vector, specifying variables/fields to be selected}

\item{sort}{a named vector, specifying sort order in the form "variable" = "asc"}

\item{verbose}{a boolean, specifying whether to print verbose messages}
}
\value{
JSON data object i.e. a list
}
\description{
\code{get_JSON_doc} will return infomation about the requested resource. Ideally, will be just used internally.
}
\examples{
\dontrun{
library(RCurl)
library(RJSONIO)
# Return 100 elements from a hotels data resource
JSON_doc = get_JSON_doc(link="http://data.gov.in/api/datastore/resource.json?",
   res_id="0749068c-a590-4a07-a571-e9df5dddcc8a",
   offset=0,
   no_elements=100)
}
}
\keyword{Name}

