% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogdindiar.R
\docType{package}
\name{ogdindiar}
\alias{ogdindiar}
\alias{ogdindiar-package}
\title{ogdindiar: Provides API access to selected datasets on Open Government Data - India Portal.}
\description{
The ogdindiar package provides three categories of important functions:
Downloading entire datasets,
Downloading specific elelments based on certain conditions, and
Search for data sets.
}
\section{ogdindiar functions}{

fetch_data
search_datasets
}

